pulled settings out of modal file into separate file
